Fix a loss of precision when assigning palette color in VGA other
The return value from palette_num() is assigned into 8-bit
variables, therefore it only needs to return an 8-bit value.
Likewise, it only returns numbers between zero and three,
so there is no concern about rollover.

This return-type change fixes two loss of precision warnings:

implicit conversion loses integer precision: 'int' to 'Bit8u' (aka 'unsigned char')
    herc_pal = palette_num(colour);
                         ~ ^~~~~~~~~~~~~~~~~~~
implicit conversion loses integer precision: 'int' to 'Bit8u' (aka 'unsigned char')
    mono_cga_pal = palette_num(colour);

+++ src/hardware/vga_other.cpp
@@ -816,7 +816,7 @@ static void write_pcjr(Bitu port,Bitu val,Bitu /*iolen*/) {
 	}
 }
 
-static int palette_num(const char *colour)
+static uint8_t palette_num(const char *colour)
 {
 	if (strcasecmp(colour, "green") == 0)
 		return 0;
